rename access_mgr to accessmanager
