+++ packages/chains/accessmanager/access_manager.go
@@ -1,7 +1,7 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-package access_mgr
+package accessmanager
 
 import (
 	"context"
@@ -11,7 +11,7 @@ import (
 
 	"github.com/iotaledger/hive.go/log"
 
-	"github.com/iotaledger/wasp/packages/chains/access_mgr/am_dist"
+	"github.com/iotaledger/wasp/packages/chains/accessmanager/am_dist"
 	"github.com/iotaledger/wasp/packages/cryptolib"
 	"github.com/iotaledger/wasp/packages/gpa"
 	"github.com/iotaledger/wasp/packages/isc"
