+++ packages/chains/accessmanager/am_dist/WaspChainAccessNodesV4_N2C1L3R0.cfg
