+++ packages/chains/accessmanager/am_dist/WaspChainAccessNodesV4_N2C1L4R0.cfg
