+++ packages/chains/accessmanager/am_dist/access_manager_dist_test.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/iotaledger/wasp/packages/chains/access_mgr/am_dist"
+	"github.com/iotaledger/wasp/packages/chains/accessmanager/am_dist"
 	"github.com/iotaledger/wasp/packages/cryptolib"
 	"github.com/iotaledger/wasp/packages/gpa"
 	"github.com/iotaledger/wasp/packages/isc"
