+++ packages/chains/accessmanager/am_dist/access_manager_rapid_test.go
@@ -12,7 +12,7 @@ import (
 	"pgregory.net/rapid"
 
 	"github.com/iotaledger/hive.go/log"
-	"github.com/iotaledger/wasp/packages/chains/access_mgr/am_dist"
+	"github.com/iotaledger/wasp/packages/chains/accessmanager/am_dist"
 	"github.com/iotaledger/wasp/packages/cryptolib"
 	"github.com/iotaledger/wasp/packages/gpa"
 	"github.com/iotaledger/wasp/packages/isc"
